backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.35.0
sentence-transformers==3.0.1
pdfplumber==0.11.1
python-docx==1.1.2
httpx==0.27.0
beautifulsoup4==4.12.3
radon==6.0.1
gitpython==3.1.43
celery==5.4.0
redis==5.0.7
pgvector==0.3.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
tenacity==8.4.2
